feat: 길드 contributions이 잘못 변경되는 버그를 수정한다

+++ src/main/kotlin/org/gitanimals/guild/saga/UpdateGuildContributionSagaHandler.kt
@@ -20,7 +20,7 @@ class UpdateGuildContributionSagaHandler(
 
         guildService.updateContribution(
             username = userContributionUpdated.username,
-            contributions = userContributionUpdated.contributions.toLong(),
+            contributions = userContributionUpdated.contributions,
         )
     }
 }
